+++ LootLogger"Bushra_Oudah-Itr3 "Ch 12/LootLogger/Support/AppDelegate.swift
