+++ LootLogger"Bushra_Oudah-Itr3 "Ch 12/LootLogger/Support/Assets.xcassets/AppIcon.appiconset/Contents.json
